Fix view losing sync after animation interruption
Example: calling headRight quickly in succession, such as in a for-loop,
would move the tape data structure correctly, but would move the graphic
by only one cell.

Fixed by keeping DOM in sync regardless of transitions:
tag exiting nodes as '.exiting', and ensure the pre-condition that
all exiting nodes are removed before the main logic of
write and headRight/Left.

Fixed flickering when writing over with the same symbol,
by not animating when the symbol is the same.